Prep for separate RO and RW LDAP hosts
The list of RW_HOSTS will be only be used when LDAP is bound to a specific
user id.  The RO_HOSTS will be used for all other operations.

Fix a problem where the configuration method was picking up the set of LDAP
servers from /etc/[open]ldap/ldap.config instead of the current set of
servers.