rough in an empty svn test directory
Now all that is needed is some test data.

Sebb: Feel free to adjust the name of the directory as you see fit.

Note: since the directory in question is read-only, checking in dummy
data into git should be sufficient.  In the event that updating of svn
data is needed for a test, see the calls to `svnadmin` in
whimsy/board/agenda/Rakefile to see how a local repository can be created,
populated with data, and have a working directory checked out locally.